selstagram2/requirements.txt
amqp==2.1.4
Babel==2.4.0
billiard==3.5.0.2
celery==4.0.2
Django==1.11
django-celery-beat==1.0.1
django-celery-results==1.0.1
django-model-utils==3.0.0
djangorestframework==3.6.2
drf-nested-routers==0.90.0
factory-boy==2.8.1
flower==0.9.1
instaLooter==0.8.2
kombu==4.0.2
munch==2.1.1
psycopg2==2.7.1
pytz==2017.2
SQLAlchemy==1.1.9
tornado==4.2
vine==1.1.3
